Add TeamBoardPolicy

+++ app/Http/Controllers/Api/TeamBoardController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\Team\Board\StoreBoardRequest;
 use App\Http\Requests\Team\Board\UpdateBoardRequest;
 use App\Models\Board;
 use App\Models\Team;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 
 class TeamBoardController extends Controller
@@ -14,9 +15,11 @@ class TeamBoardController extends Controller
     /**
      * @param Team $team
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function index(Team $team)
     {
+        $this->authorize('view', $team);
         return $this->successResponse($team->boards, 'Auth user team boards', 200);
     }
 
@@ -24,9 +27,11 @@ class TeamBoardController extends Controller
      * @param StoreBoardRequest $request
      * @param Team $team
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function store(StoreBoardRequest $request, Team $team)
     {
+        $this->authorize('create', $team);
         try {
             $board = $team->boards()->create($request->validated());
         } catch (\Exception $exception) {
@@ -39,9 +44,11 @@ class TeamBoardController extends Controller
      * @param Team $team
      * @param Board $board
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function show(Team $team, Board $board)
     {
+        $this->authorize('viewAdvanced', [$team, $board]);
         return $this->successResponse($board, 'Board detail', 200);
     }
 
@@ -50,9 +57,11 @@ class TeamBoardController extends Controller
      * @param Team $team
      * @param Board $board
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function update(UpdateBoardRequest $request, Team $team, Board $board)
     {
+        $this->authorize('update', [$team, $board]);
         try {
             $board->update($request->validated());
         } catch (\Exception $exception) {
@@ -68,6 +77,7 @@ class TeamBoardController extends Controller
      */
     public function destroy(Team $team, Board $board)
     {
+        $this->authorize('delete', [$team, $board]);
         try {
             $board->delete();
         } catch (\Exception $exception) {
